refactor(api): archive and unarchive route migration to src
