+++ api/lib/application/campaigns/index.js
@@ -115,41 +115,6 @@ const register = async function (server) {
       },
     },
     {
-      method: 'PUT',
-      path: '/api/campaigns/{id}/archive',
-      config: {
-        pre: [{ method: securityPreHandlers.checkAuthorizationToManageCampaign }],
-        validate: {
-          params: Joi.object({
-            id: identifiersType.campaignId,
-          }),
-        },
-        handler: campaignController.archiveCampaign,
-        tags: ['api'],
-        notes: [
-          '- **Cette route est restreinte aux utilisateurs authentifiés**\n' + "- Archivage d'une campagne par son id",
-        ],
-      },
-    },
-    {
-      method: 'DELETE',
-      path: '/api/campaigns/{id}/archive',
-      config: {
-        pre: [{ method: securityPreHandlers.checkAuthorizationToManageCampaign }],
-        validate: {
-          params: Joi.object({
-            id: identifiersType.campaignId,
-          }),
-        },
-        handler: campaignController.unarchiveCampaign,
-        tags: ['api'],
-        notes: [
-          '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
-            "- Désarchivage d'une campagne par son id",
-        ],
-      },
-    },
-    {
       method: 'GET',
       path: '/api/campaigns/{id}/divisions',
       config: {
