+++ api/src/prescription/campaign/application/campaign-adminstration-controller.js
@@ -85,6 +85,20 @@ const updateCampaignCode = async function (request, h) {
   return h.response(null).code(204);
 };
 
+const archiveCampaign = function (request, h, dependencies = { campaignReportSerializer }) {
+  const { userId } = request.auth.credentials;
+  const campaignId = request.params.id;
+
+  return usecases.archiveCampaign({ userId, campaignId }).then(dependencies.campaignReportSerializer.serialize);
+};
+
+const unarchiveCampaign = function (request, h, dependencies = { campaignReportSerializer }) {
+  const { userId } = request.auth.credentials;
+  const campaignId = request.params.id;
+
+  return usecases.unarchiveCampaign({ userId, campaignId }).then(dependencies.campaignReportSerializer.serialize);
+};
+
 const campaignAdministrationController = {
   save,
   update,
@@ -92,6 +106,8 @@ const campaignAdministrationController = {
   swapCampaignCodes,
   updateCampaignDetails,
   updateCampaignCode,
+  archiveCampaign,
+  unarchiveCampaign,
 };
 
 export { campaignAdministrationController };
