+++ api/src/prescription/campaign/domain/errors.js
@@ -40,7 +40,14 @@ class CampaignCodeFormatError extends DomainError {
   }
 }
 
+class ArchivedCampaignError extends DomainError {
+  constructor(message = 'Cette campagne est déjà archivée.') {
+    super(message);
+  }
+}
+
 export {
+  ArchivedCampaignError,
   CampaignCodeFormatError,
   CampaignUniqueCodeError,
   IsForAbsoluteNoviceUpdateError,
