+++ api/src/prescription/campaign/domain/models/CampaignForArchiving.js
@@ -1,4 +1,5 @@
-import { ArchivedCampaignError, ObjectValidationError } from '../errors.js';
+import { ObjectValidationError } from '../../../../../lib/domain/errors.js';
+import { ArchivedCampaignError } from '../errors.js';
 
 class CampaignForArchiving {
   constructor({ id, code, archivedAt, archivedBy } = {}) {
