+++ api/src/prescription/campaign/domain/usecases/archive-campaign.js
