+++ api/src/prescription/campaign/domain/usecases/index.js
@@ -19,6 +19,7 @@ import { importNamedExportsFromDirectory } from '../../../../shared/infrastructu
 import * as campaignAdministrationRepository from '../../infrastructure/repositories/campaign-administration-repository.js';
 import * as campaignAssessmentParticipationResultListRepository from '../../infrastructure/repositories/campaign-assessment-participation-result-list-repository.js';
 import * as campaignCreatorRepository from '../../infrastructure/repositories/campaign-creator-repository.js';
+import * as campaignForArchivingRepository from '../../infrastructure/repositories/campaign-for-archiving-repository.js';
 // TODO : clean with campaign administration ( similar get with a lot difference)
 import * as campaignManagementRepository from '../../infrastructure/repositories/campaign-management-repository.js';
 import { campaignParticipantActivityRepository } from '../../infrastructure/repositories/campaign-participant-activity-repository.js';
@@ -56,6 +57,7 @@ const dependencies = {
   stageCollectionRepository,
   targetProfileRepository, // TODO
   userRepository,
+  campaignForArchivingRepository,
 };
 
 const path = dirname(fileURLToPath(import.meta.url));
