+++ api/tests/prescription/campaign/unit/application/campaign-administration-route_test.js
@@ -364,4 +364,32 @@ describe('Unit | Application | Router | campaign-administration-router ', functi
       expect(response.statusCode).to.equal(409);
     });
   });
+
+  describe('PUT /api/campaigns/{id}/archive', function () {
+    it('should return 400 with an invalid campaign id', async function () {
+      // given
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+
+      // when
+      const response = await httpTestServer.request('PUT', '/api/campaigns/invalid/archive');
+
+      // then
+      expect(response.statusCode).to.equal(400);
+    });
+  });
+
+  describe('DELETE /api/campaigns/{id}/archive', function () {
+    it('should return 400 with an invalid campaign id', async function () {
+      // given
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+
+      // when
+      const response = await httpTestServer.request('DELETE', '/api/campaigns/invalid/archive');
+
+      // then
+      expect(response.statusCode).to.equal(400);
+    });
+  });
 });
