+++ api/tests/prescription/campaign/unit/domain/models/CampaignForArchiving_test.js
@@ -1,6 +1,7 @@
-import { ArchivedCampaignError, ObjectValidationError } from '../../../../lib/domain/errors.js';
-import { CampaignForArchiving as Campaign } from '../../../../lib/domain/models/CampaignForArchiving.js';
-import { catchErr, expect } from '../../../test-helper.js';
+import { ObjectValidationError } from '../../../../../../lib/domain/errors.js';
+import { ArchivedCampaignError } from '../../../../../../src/prescription/campaign/domain/errors.js';
+import { CampaignForArchiving as Campaign } from '../../../../../../src/prescription/campaign/domain/models/CampaignForArchiving.js';
+import { catchErr, expect } from '../../../../../test-helper.js';
 
 describe('Unit | Domain | Models | CampaignForArchiving', function () {
   describe('#archive', function () {
