+++ api/tests/prescription/campaign/unit/domain/usecases/archive-campaign_test.js
@@ -1,6 +1,6 @@
-import { CampaignForArchiving as Campaign } from '../../../../lib/domain/models/CampaignForArchiving.js';
-import { archiveCampaign } from '../../../../lib/domain/usecases/archive-campaign.js';
-import { expect, sinon } from '../../../test-helper.js';
+import { CampaignForArchiving as Campaign } from '../../../../../../src/prescription/campaign/domain/models/CampaignForArchiving.js';
+import { archiveCampaign } from '../../../../../../src/prescription/campaign/domain/usecases/archive-campaign.js';
+import { expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | UseCase | archive-campaign', function () {
   let clock;
