+++ api/tests/prescription/campaign/unit/domain/usecases/unarchive-campaign_test.js
@@ -1,6 +1,6 @@
-import { CampaignForArchiving as Campaign } from '../../../../lib/domain/models/CampaignForArchiving.js';
-import { unarchiveCampaign } from '../../../../lib/domain/usecases/unarchive-campaign.js';
-import { expect, sinon } from '../../../test-helper.js';
+import { CampaignForArchiving as Campaign } from '../../../../../../src/prescription/campaign/domain/models/CampaignForArchiving.js';
+import { unarchiveCampaign } from '../../../../../../src/prescription/campaign/domain/usecases/unarchive-campaign.js';
+import { expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | UseCase | unarchive-campaign', function () {
   let campaignForArchivingRepository;
