+++ api/tests/unit/application/campaign/campaign-controller_test.js
@@ -165,81 +165,4 @@ describe('Unit | Application | Controller | Campaign', function () {
       expect(errorCatched).to.be.instanceof(UserNotAuthorizedToAccessEntityError);
     });
   });
-
-  describe('#archiveCampaign', function () {
-    let updatedCampaign;
-    let serializedCampaign;
-
-    const campaignId = 1;
-    const userId = 1;
-    let campaignReportSerializerStub;
-    beforeEach(function () {
-      sinon.stub(usecases, 'archiveCampaign');
-      campaignReportSerializerStub = {
-        serialize: sinon.stub(),
-      };
-      updatedCampaign = Symbol('updated campaign');
-      serializedCampaign = Symbol('serialized campaign');
-    });
-
-    it('should return the updated campaign properly serialized', async function () {
-      // given
-      usecases.archiveCampaign.withArgs({ userId, campaignId }).resolves(updatedCampaign);
-      campaignReportSerializerStub.serialize.withArgs(updatedCampaign).returns(serializedCampaign);
-
-      // when
-      const response = await campaignController.archiveCampaign(
-        {
-          params: { id: campaignId },
-          auth: {
-            credentials: { userId },
-          },
-        },
-        hFake,
-        { campaignReportSerializer: campaignReportSerializerStub },
-      );
-
-      // then
-      expect(response).to.be.equal(serializedCampaign);
-    });
-  });
-
-  describe('#unarchiveCampaign', function () {
-    let updatedCampaign;
-    let serializedCampaign;
-
-    const campaignId = 1;
-    const userId = 1;
-    let campaignReportSerializerStub;
-    beforeEach(function () {
-      sinon.stub(usecases, 'unarchiveCampaign');
-
-      campaignReportSerializerStub = {
-        serialize: sinon.stub(),
-      };
-      updatedCampaign = Symbol('updated campaign');
-      serializedCampaign = Symbol('serialized campaign');
-    });
-
-    it('should return the updated campaign properly serialized', async function () {
-      // given
-      usecases.unarchiveCampaign.withArgs({ userId, campaignId }).resolves(updatedCampaign);
-      campaignReportSerializerStub.serialize.withArgs(updatedCampaign).returns(serializedCampaign);
-
-      // when
-      const response = await campaignController.unarchiveCampaign(
-        {
-          params: { id: campaignId },
-          auth: {
-            credentials: { userId },
-          },
-        },
-        hFake,
-        { campaignReportSerializer: campaignReportSerializerStub },
-      );
-
-      // then
-      expect(response).to.be.equal(serializedCampaign);
-    });
-  });
 });
