+++ api/tests/unit/domain/usecases/archive-campaign-from-campaign-code_test.js
@@ -1,5 +1,5 @@
-import { CampaignForArchiving as Campaign } from '../../../../lib/domain/models/CampaignForArchiving.js';
 import { archiveCampaignFromCampaignCode } from '../../../../lib/domain/usecases/archive-campaign-from-campaign-code.js';
+import { CampaignForArchiving as Campaign } from '../../../../src/prescription/campaign/domain/models/CampaignForArchiving.js';
 import { expect, sinon } from '../../../test-helper.js';
 
 describe('Unit | UseCase | archive-campaign', function () {
